+++ dvha/options.py
@@ -219,6 +219,16 @@ class DefaultOptions:
 
         self.SHOW_NEW_PTV_CALC_WARNING = True
 
+        self.GET_DVH_KWARGS = {'calculate_full_volume': True,
+                               'use_structure_extents': False,
+                               'interpolation_resolution': None,
+                               'interpolation_segments_between_planes': 0,
+                               'memmap_rtdose': False}
+        self.DVH_SMALL_VOLUME_THRESHOLD = 10  # compute high resolution DVH if volume less than this (cc)
+        self.DVH_HIGH_RESOLUTION_FACTOR = 8  # Must be a factor of (2^n)/4, where n is an integer
+        self.DVH_HIGH_RESOLUTION_FACTOR_OPTIONS = ['4', '8', '16', '32']
+        self.DVH_HIGH_RESOLUTION_SEGMENTS_BETWEEN = 2  # Must be int
+
 
 class Options(DefaultOptions):
     def __init__(self):